Fix Poisson initialization precision

+++ math/probability/poisson.py
@@ -26,7 +26,7 @@ class Poisson:
                 raise TypeError("data must be a list")
             if len(data) < 2:
                 raise ValueError("data must contain multiple values")
-            self.lambtha = float(sum(data) / len(data))
+            self.lambtha = sum(data) / len(data)
 
     def pmf(self, k):
         """Calculate PMF value for k successes.
